debt - layering fixes
