+++ src/vs/platform/diagnostics/electron-sandbox/diagnosticsService.ts
